SwiftExtract: move Foundation overlay logic into SwiftAnalyzer+Foundation.swift
`visitFoundationDeclsIfNeeded` is a self-contained ~65-line block
that overlays Foundation/FoundationEssentials known types when the
analyzed code references any of them. It belongs next to the rest of
the analyzer's Foundation knowledge but doesn't need to live in the
main `SwiftAnalyzer.swift` file. Move it into a sibling
`SwiftAnalyzer+Foundation.swift` (still an extension on
`SwiftAnalyzer`).

Visibility goes from `private` to `internal` (no modifier) since
its caller `analyze(...)` lives in the original file. The `isUsing`
helper it depends on stays where it is and remains internal.
